move pre-exisiting helper functions to utils
